fix: Kotlin 컴파일 오류 6건 수정
- RiskCalculator: ${finalScore}점 템플릿 문자열 수정 (한글 식별자 오류)
- CalculateRiskUseCase: sumOf 람다에 .toDouble() 추가 (타입 추론 실패)
- DeviceListItem: SUSPECTED_CAMERA → IP_CAMERA (존재하지 않는 enum)
- IrCameraScreen: LocalLifecycleOwner import 추가
- LensFinderScreen: LocalLifecycleOwner import 추가
- OnboardingScreen: @OptIn(ExperimentalFoundationApi::class) 추가

+++ app/src/main/java/com/searcam/domain/usecase/CalculateRiskUseCase.kt
@@ -51,7 +51,7 @@ class CalculateRiskUseCase {
             } else {
                 layerResult.layerType.weightNoWifi
             }
-            weight * layerResult.score
+            (weight * layerResult.score).toDouble()
         }
 
         val positiveCount = completedLayers.count { it.isPositive }
